sched/core: Fix build paravirt build on arm and arm64
Commit 004172bdad64 ("sched/core: Remove unnecessary #include headers")
removed the inclusion of asm/paravirt.h which is used to get
declarations of paravirt_steal_rq_enabled and paravirt_steal_clock.

It is implicitly included on x86 but not on arm and arm64 breaking the
build if paravirtualization is used.  Since things from that header are
used directly fix the build by putting the direct inclusion back.

Signed-off-by: Mark Brown <broonie@kernel.org>
Change-Id: I9ef99550a4ed44366f6f8affeaf119a0339f9977
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>