ANDROID: clock_gettime(CLOCK_BOOTTIME,) slows down >20x
clock_gettime(CLOCK_BOOTTIME,) slows down after significant
accumulation of suspend time creating a large offset between it and
CLOCK_MONOTONIC time.  The __iter_div_u64_rem() is only for the usage
of adding a few second+nanosecond times and saving cycles on more
expensive remainder and division operations, but iterates one second
at a time which quickly goes out of scale in CLOCK_BOOTTIME's case
since it was specified as nanoseconds only.

The fix is to split off seconds from the boot time and cap the
nanoseconds so that __iter_div_u64_rem does not iterate.

 Conflicts:
	arch/arm64/kernel/vdso.c

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Change-Id: I75c8722bb434c2c7e1b9a1eb066e53918755ecab
Bug: 72406285
Signed-off-by: Albert I <kras@raphielgang.org>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>