thermal: msm_lmh_dcvs: Never initialize LMH TZ more than once
The LMH DCVS initialization consists in enabling the LMH thermal,
outer loop, reliability and BCL algorithms on a per-cluster
basis, but the command that kickstarts it is "PROFILE_CHANGE"
with parameter '1' sent to the TZ LMH Service.

Sending it more than one time is rendundant, but the critical
issue is that - if we send it more than once - the CPUs will
deadlock!

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: I7c5d7e23bc1b1fa22fbef622f1b13245f1b0e0e3
Signed-off-by: TogoFire <togofire@mailfence.com>