msm: vidc: Fix broken debugfs creation error checks and error paths
The debugfs helper functions do not always return NULL when they fail;
instead, they can return an error number casted as a pointer, so that their
users have the option to determine the exact cause of failure.

Use the IS_ERR_OR_NULL() helper when checking for debugfs errors to fix the
error checks.

This fixes the kernel panic on boot when debugfs is disabled.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I8342443a5e30c580181e4e56c332bdc2358a5d99
Signed-off-by: Lau <laststandrighthere@gmail.com>