msm: kgsl: improve memory stats maintenance
While freeing memory, page_alloc memstat wrongly shows
that the memory is freed while the process is still ongoing.
While allocating memory if something goes wrong during the
process, the memory allocated till that point is not
reflected in page_alloc. Change the point of updation of
page_alloc. Also add two stats, one to track allocation in
progress and the other to track free in progress.

Signed-off-by: Puranam V G Tejaswi <pvgtejas@codeaurora.org>
Change-Id: Ic637de02cb188a5de4f9c3703897193a099be570
Signed-off-by: Lau <laststandrighthere@gmail.com>