prlmk: Use vmpressure to stop reclaims/kills early
This patch attempts to kill/reclaim even fewer tasks, and maintain a
stable memory pressure in the system at the same time.We use vmpressure
to determine when the memory pressure is lower than pressure_max, and
then we stop killing/reclaiming tasks.

Also, we try to sort the tasks so that the largest tasks gets reclaimed
first.This should reduce the memory pressure faster.

Granted, we may be triggering the reclaim/kill workqueue based on
a memory spike, but the pressure checks as explained above should
stop unnecessary kills.


















-e 






























Change-Id: I4f4c2469113d6259814ce899895b1b1c57d87a22