thermal_core: Use notify ops when CONFIG_ACPI_THERMAL=y
Commit [1] has removed notifys user for the acpi thermal driver.
Since only acpi thermal driver uses notify ops currently, put a
guard over the notify ops at thermal_core to not generate codes
for it when the driver is not used. Running a conditional check
for a driver that we don't use is of no benefit.

[1]: https://github.com/Atom-X-Devs/android_kernel_xiaomi_scarlet/commit/04f111130e9afa41c10d7bcec14e00e3be8b6977

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: Iec8a0342d9aea5b7d9f3582248386190036640c2
Signed-off-by: TogoFire <togofire@mailfence.com>