fingerprint: fpc: Fix loop counter range for pinctrl checkup
The "for" loop counter was setting its range till the array size of
pinctrl_state. But pinctrl_state is a pointer for the array size of
pctl_names. Using ARRAY_SIZE() twice is incorrect. So, just replace
pinctrl_state with pctl_names for the loop counter range for the
pinctrl checkup.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: I6874fc6a9c62f3ead2aa8e5b90207e6460bcc6b3
Signed-off-by: TogoFire <togofire@mailfence.com>